requirements.txt
pytesseract
Pillow
numpy
opencv-python
pystray
